requirements.txt
streamlit
python_dotenv
google-generativeai
fpdf